// src/themes/types/TypeMicroInteractions.d.ts
export interface TypeMicroInteractions {
    card: {
        enter: string;
        accentColor: string;
        accentTransition: string;
    };

    menuItem: {
        hoverBg: string;
        hoverTransform: string;
        transition: string;
    };

    rowTable: {
        hoverBg: string;
        hoverTransform: string;
        hoverBoxShadow: string
        transition: string;
    };

    clickElement: {
        hoverCursor: string;
        hoverColor: string;
        transition: string;
    }
}